Split of hotswap_test.go into baseline and extended
